+++ src/beefcommands/cogs/utilities_cog.py
@@ -3,7 +3,8 @@ from discord.ext import commands
 from beefcommands.utilities.ccping import ccping
 from beefutilities.guilds import set_info, set_logs, set_quotes
 from beefcommands.utilities.help import help
-from beefcommands.utilities.music_player.yt_link_parser import get_audio_link
+from beefcommands.utilities.music_player.youtube.get_yt_audio import get_audio_link
+from beefcommands.utilities.music_player.link_parser import link_validation
 
 class UtilitiesCog(commands.Cog):
     def __init__(self, bot):
@@ -15,7 +16,12 @@ class UtilitiesCog(commands.Cog):
     
     @discord.app_commands.command(name="ccinfo", description="returns current info about CCServer")
     async def ccinfo(self, interaction: discord.Interaction):
-        await get_audio_link(interaction, "aurora - runaway")
+        url = "https://www.twitch.tv/variscites"
+        type = await link_validation(url)
+        if type != "invalid":
+            print(await get_audio_link(interaction, url))
+        else:
+            print("invalid link")
     
     @discord.app_commands.command(name="set_logs_channel", description="where should i spew...?")
     async def set_log_channel(self, interacton: discord.Interaction):
