+++ src/beefcommands/utilities/music_player/link_parser.py
@@ -1,5 +1,7 @@
 import re
 import discord
+from beefcommands.utilities.music_player.youtube.yt_query_constructor import get_metadata_yt, yt_search_term
+from beefcommands.utilities.music_player.spotify.spotify_link_parser import spotify_link_parser
 
 async def link_validation(url: str):
     print("validating the link")
@@ -8,17 +10,15 @@ async def link_validation(url: str):
     search_pattern = re.compile(r'^(?!https?://).+')
 
     if youtube_pattern.match(url):
-        return True, "youtube"
+        return "youtube"
     elif spotify_pattern.match(url):
-        return True, "spotify"
+        return "spotify"
     elif search_pattern.match(url):
-        return True, "search"
+        return "search"
     else:
-        return False, "invalid"
+        return "invalid"
 
-async def link_parser(interaction: discord.Interaction, url: str):
-    (valid, type) = await link_validation(url)
-    
+async def media_source(interaction: discord.Interaction, url: str):
     if type == 'youtube':
         print("youtube link, retrieving metadata")
         return await get_metadata_yt(interaction, url)
@@ -27,9 +27,8 @@ async def link_parser(interaction: discord.Interaction, url: str):
         return await spotify_link_parser(interaction, url)
     elif type == 'search':
         print("search term, retrieving yt link")        
-        return await yt_search(interaction, url)
+        return await yt_search_term(interaction, url)
+    else:
+        return "invalid"
+        
 
-async def get_audio_link(interaction: discord.Interaction, url: str):
-    print(f"putting {url} throught the link parser...")
-    metadata = await link_parser(interaction, url)
-    return metadata["url"]
